Data_Science/Langchain_Family/pyproject.toml
[project]
name = "langchain-llm"
version = "0.1.0"
description = ""
authors = [
    {name = "KevBoyz",email = "kevboyz@pm.me"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.59,<0.4.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "langchain-google-firestore (>=0.5.0,<0.6.0)",
    "chromadb (>=1.0.8,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "firecrawl-py (>=2.5.4,<3.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "assemblyai (>=0.40.2,<0.41.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "qdrant-client (>=1.14.2,<2.0.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

